+++ test/fixture/entry_point.py
@@ -1,0 +1,1 @@
+print("Executing entry point script")